requirements.txt
tensorflow==1.11.0
wordcloud==1.5.0
matplotlib==2.2.4
Flask==0.12
google_auth_oauthlib==0.3.0
pandas==0.23.4
en_core_web_sm==2.1.0
numpy==1.15.4
spacy==2.1.3
nltk==3.4.5
protobuf==3.7.1
google_api_python_client==1.7.8
